catch2: change CMake build type to Release
Catch2 version 3 builds as a static library. It is not necessary
to build Catch2 in debug mode in order to run tests on the target.
Switch build type to Release to follow general buildroot practice
and name static libraries generated (Catch2Main.a, Catch2.a) match
those used on standard linux distros.